requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
distlib==0.3.7
filelock==3.12.2
Flask==2.3.2
Flask-Session==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
platformdirs==3.10.0
protobuf==4.21.12
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3
virtualenv==20.24.2
Werkzeug==2.3.6
